+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- Logout button. #3063
 - Add an option to the Hadoop exploiter to try all discovered HTTP ports. #2136
 - `GET /api/agent-otp`. #3076
 - `POST /api/agent-otp-login` endpoint. #3076